breaking portfolio into 2 sub-components
